pyproject.toml
[project]
name = "engenharia-de-ia-com-python"
version = "0.1.0"
description = ""
authors = [
    {name = "Anwar Hermuche",email = "anwarhermuche2@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy (>=2.4.1,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
]

[tool.poetry]
packages = [{include = "engenharia_de_ia_com_python", from = "."}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
